+++ src/takajo.nim
@@ -1,7 +1,6 @@
 import algorithm
 import cligen
 import json
-import math
 import nancy
 import puppy
 import re
@@ -11,14 +10,19 @@ import strformat
 import strutils
 import tables
 import terminal
-import termstyle
 import times
 import threadpool
 import uri
 import os
 import std/enumerate
+import std/xmlparser
+import std/xmltree
 import suru
 import takajopkg/general
+import takajopkg/stackUtil
+import takajopkg/takajoTerminal
+import takajopkg/ttpResult
+import takajopkg/vtResult
 include takajopkg/extractScriptblocks
 include takajopkg/listDomains
 include takajopkg/listIpAddresses
@@ -26,18 +30,25 @@ include takajopkg/listUndetectedEvtxFiles
 include takajopkg/listUnusedRules
 include takajopkg/splitCsvTimeline
 include takajopkg/splitJsonTimeline
+include takajopkg/stackCmdlines
+include takajopkg/stackDNS
 include takajopkg/stackLogons
+include takajopkg/stackProcesses
+include takajopkg/stackServices
+include takajopkg/stackTasks
 include takajopkg/listHashes
 include takajopkg/sysmonProcessTree
 include takajopkg/timelineLogon
 include takajopkg/timelinePartitionDiagnostic
 include takajopkg/timelineSuspiciousProcesses
+include takajopkg/timelineTasks
 include takajopkg/ttpSummary
 include takajopkg/ttpVisualize
 include takajopkg/ttpVisualizeSigma
 include takajopkg/vtDomainLookup
 include takajopkg/vtIpLookup
 include takajopkg/vtHashLookup
+
 
 when isMainModule:
     clCfg.version = "2.4.0-dev"
@@ -49,12 +60,18 @@ when isMainModule:
     const example_list_unused_rules = "  list-unused-rules -t ../hayabusa/timeline.csv -r ../hayabusa/rules\p"
     const example_split_csv_timeline = "  split-csv-timeline -t ../hayabusa/timeline.csv [--makeMultiline] -o case-1-csv\p"
     const example_split_json_timeline = "  split-json-timeline -t ../hayabusa/timeline.jsonl -o case-1-json\p"
+    const example_stack_cmdlines = "  stack-cmdlines -t ../hayabusa/timeline.jsonl [--level low] -o cmdlines.csv\p"
+    const example_stack_dns = "  stack-dns -t ../hayabusa/timeline.jsonl [--level infomational]  -o dns.csv\p"
     const example_stack_logons = "  stack-logons -t ../hayabusa/timeline.jsonl -o logons.csv\p"
+    const example_stack_services  = "  stack-services -t ../hayabusa/timeline.jsonl [--level infomational] -o services.csv\p"
+    const example_stack_tasks = "  stack-tasks -t ../hayabusa/timeline.jsonl [--level infomational] -o tasks.csv\p"
+    const example_stack_processes = "  stack-processes -t ../hayabusa/timeline.jsonl [--level low] -o processes.csv\p"
     const example_list_hashes = "  list-hashes -t ../hayabusa/case-1.jsonl -o case-1\p"
     const example_sysmon_process_tree = "  sysmon-process-tree -t ../hayabusa/timeline.jsonl -p <Process GUID> [-o process-tree.txt]\p"
     const example_timeline_logon = "  timeline-logon -t ../hayabusa/timeline.jsonl -o logon-timeline.csv\p"
     const example_timeline_partition_diagnostic = "  timeline-partition-diagnostic -t ../hayabusa/timeline.jsonl -o partition-diagnostic-timeline.csv\p"
     const example_timeline_suspicious_processes = "  timeline-suspicious-processes -t ../hayabusa/timeline.jsonl [--level medium] [-o suspicious-processes.csv]\p"
+    const example_timeline_tasks = "  timeline-tasks -t ../hayabusa/timeline.jsonl -o task-timeline.csv\p"
     const example_vt_domain_lookup = "  vt-domain-lookup  -a <API-KEY> --domainList domains.txt -r 1000 -o results.csv --jsonOutput responses.json\p"
     const example_ttp_summary = "  ttp-summary -t ../hayabusa/timeline.jsonl -o ttp-summary.csv\p"
     const example_ttp_visualize = "  ttp-visualize -t ../hayabusa/timeline.jsonl -o mitre-ttp-heatmap.json\p"
@@ -62,11 +79,13 @@ when isMainModule:
     const example_vt_hash_lookup = "  vt-hash-lookup -a <API-KEY> --hashList case-1-MD5-hashes.txt -r 1000 -o results.csv --jsonOutput responses.json\p"
     const example_vt_ip_lookup = "  vt-ip-lookup -a <API-KEY> --ipList ipAddresses.txt -r 1000 -o results.csv --jsonOutput responses.json\p"
 
-    clCfg.useMulti = "Version: 2.3.1 Year Of The Dragon Release\pUsage: takajo.exe <COMMAND>\p\pCommands:\p$subcmds\pCommand help: $command help <COMMAND>\p\p" &
+    clCfg.useMulti = "Version: 2.4.0 Dev Build\pUsage: takajo.exe <COMMAND>\p\pCommands:\p$subcmds\pCommand help: $command help <COMMAND>\p\p" &
         examples & example_extract_scriptblocks &
         example_list_domains & example_list_hashes & example_list_ip_addresses & example_list_undetected_evtx & example_list_unused_rules &
-        example_split_csv_timeline & example_split_json_timeline & example_stack_logons & example_sysmon_process_tree &
-        example_timeline_logon & example_timeline_partition_diagnostic & example_timeline_suspicious_processes &
+        example_split_csv_timeline & example_split_json_timeline &
+        example_stack_cmdlines & example_stack_dns & example_stack_logons & example_stack_processes & example_stack_services & example_stack_tasks &
+        example_sysmon_process_tree &
+        example_timeline_logon & example_timeline_partition_diagnostic & example_timeline_suspicious_processes & example_timeline_tasks &
         example_ttp_summary & example_ttp_visualize & example_ttp_visualize_sigma &
         example_vt_domain_lookup & example_vt_hash_lookup & example_vt_ip_lookup
 
@@ -77,7 +96,7 @@ when isMainModule:
             extractScriptblocks, cmdName = "extract-scriptblocks",
             doc = "extract and reassemble PowerShell EID 4104 script block logs",
             help = {
-                "level": "specify the minimum alert level",
+                "level": "specify the minimum alert level (default: low)",
                 "output": "output directory (default: scriptblock-logs)",
                 "quiet": "do not display the launch banner",
                 "timeline": "Hayabusa JSONL timeline (profile: any)",
@@ -166,10 +185,76 @@ when isMainModule:
             }
         ],
         [
+            stackCmdlines, cmdName = "stack-cmdlines",
+            doc = "stack executed command lines",
+            help = {
+                "level": "specify the minimum alert level (default: low)",
+                "ignoreSysmon": "exclude Sysmon 1 events",
+                "ignoreSecurity": "exclude Security 4688 events",
+                "output": "save results to a CSV file",
+                "quiet": "do not display the launch banner",
+                "timeline": "Hayabusa JSONL timeline (profile: any besides all-field-info*)",
+            },
+            short = {
+                "ignoreSysmon": 'y',
+                "ignoreSecurity": 'e'
+            }
+        ],
+        [            
+            stackDNS, cmdName = "stack-dns",
+            doc = "stack DNS queries and responses",
+            help = {
+                "level": "specify the minimum alert level (default: informational)",
+                "output": "save results to a CSV file",
+                "quiet": "do not display the launch banner",
+                "timeline": "Hayabusa JSONL timeline (profile: any besides all-field-info*)",
+            }
+        ],
+        [
             stackLogons, cmdName = "stack-logons",
             doc = "stack logons by target user, target computer, source IP address and source computer",
             help = {
                 "localSrcIpAddresses": "include results when the source IP address is local",
+                "output": "save results to a CSV file",
+                "quiet": "do not display the launch banner",
+                "timeline": "Hayabusa JSONL timeline (profile: any besides all-field-info*)",
+            }
+        ],
+        [
+            stackProcesses, cmdName = "stack-processes",
+            doc = "stack executed processes",
+            help = {
+                "level": "specify the minimum alert level (default: low)",
+                "ignoreSysmon": "exclude Sysmon 1 events",
+                "ignoreSecurity": "exclude Security 4688 events",
+                "output": "save results to a CSV file",
+                "quiet": "do not display the launch banner",
+                "timeline": "Hayabusa JSONL timeline (profile: any besides all-field-info*)",
+            },            
+            short = {
+                "ignoreSysmon": 'y',
+                "ignoreSecurity": 'e'
+            }
+        ],
+        [
+            stackServices, cmdName = "stack-services",
+            doc = "stack service names and paths",
+            help = {
+                "level": "specify the minimum alert level (default: informational)",
+                "output": "save results to a CSV file",
+                "quiet": "do not display the launch banner",
+                "timeline": "Hayabusa JSONL timeline (profile: any besides all-field-info*)",
+            },
+            short = {
+                "ignoreSystem": 'y',
+                "ignoreSecurity": 'e'
+            }
+        ],
+        [
+            stackTasks, cmdName = "stack-tasks",
+            doc = "stack new scheduled tasks",
+            help = {
+                "level": "specify the minimum alert level (default: informational)",
                 "output": "save results to a CSV file",
                 "quiet": "do not display the launch banner",
                 "timeline": "Hayabusa JSONL timeline (profile: any besides all-field-info*)",
@@ -218,6 +303,15 @@ when isMainModule:
                 "output": "save results to a CSV file",
                 "quiet": "do not display the launch banner",
                 "timeline": "Hayabusa JSONL timeline (profile: any besides all-field-info*)",
+            }
+        ],
+        [
+            timelineTasks, cmdName = "timeline-tasks",
+            doc = "create a CSV timeline of scheduled tasks",
+            help = {
+                "output": "save results to a CSV file",
+                "quiet": "do not display the launch banner",
+                "timeline": "Hayabusa JSONL timeline (profile: any)",
             }
         ],
         [